final version.

+++ src/main/java/com/khrapkov/utask/dto/response/ResponseTotalAmountDto.java
@@ -2,8 +2,10 @@ package com.khrapkov.utask.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ResponseTotalAmountDto {
     private String sender;
-    private Long totalAmount;
+    private BigDecimal totalAmount;
 }
