+++ src/main/java/com/khrapkov/utask/repository/third_db_repository/ThirdDBPaymentRepository.java
@@ -4,7 +4,9 @@ import com.khrapkov.utask.entity.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
+
 public interface ThirdDBPaymentRepository extends JpaRepository<PaymentEntity, Long> {
     @Query(" SELECT SUM(amount) FROM PaymentEntity WHERE sender = :sender")
-    Long getTotalAmountBySender(String sender);
+    BigDecimal getTotalAmountBySender(String sender);
 }
